Fix crash when person not found at the census side
If the authorization has an associated person_id, but that person is not
found on the census side, we get a crash. Not sure if this can happen,
but it's free so let's stay on the safe side.

In that admittedly weird case, user would be prompted to start census
registration over.