Connector error handling (#49)
* Fix crash when person not found at the census side

If the authorization has an associated person_id, but that person is not
found on the census side, we get a crash. Not sure if this can happen,
but it's free so let's stay on the safe side.

In that admittedly weird case, user would be prompted to start census
registration over.

* Refactor person proxy

I find it very harder to understand when a class can be instantiated by
passing different types of objects.

* Remove unnecessary check

Already done by the superclass.

* Add person to the list of context delegations

* Pass the whole person proxy as handler context

* Check actual missing person on census side

* Rename person to census_person in commands

Since there's another "Person" class, hopefully this name makes it
clearer which class it is.

* Remove dead code

* Remove complexity from forms

* Remove system identifier config

This is a decidim plugin, I don't see the configurabiltiy here.

* Move `find` to an instance method in API::Person

Since it also acts on a specific qualified id.

* Remove global id stuff

* "at census" ids are the sole responsability of API::Person since
  they're only used to connect with the API.

* "at decidm" ids are the sole responsability of API::Person.create,
  since they're only used to inform the API of the external id on
  creation.

* A bit less method chaining

* Remove complicated inheritance hierarchy in forms

* Remove more unnecessary stuff

The data handler doesn't need to be an authorization handler either.

* Rename all the things

After the previous refactorings, we need to rename all the stuff to
reconcile everything.

* Fix rubocop issue introduced by previous refactor

* Move API::Person.create to an instance method too

* Remove age validation for now

I _think_ this should be done in census.

* Minor style tweak

This is the most standard style even though rubocop is not able to
enforce it.

* Extract a `local_address?` method

Analogous to the `local_document?` one.

* Remove indirections not buying much

* Start checking create/update API status

And acting accordingly.

* Remove unnecessary validation exclusion

The document scope will always be present.

* Add missing scope_id validation

* Remove bad validation

It's incorrect and it's not actually present on the census side...

* Dinamically add API validation errors to form

* Update plugin homepages

* Use explicit parameter lists

It's more readable and maintainable in my opinion. Also, these are no
longer `AuthorizationHandler`'s so we no longer need to do the
excluding.

* Remove non-attribute validation

Validations should apply to form attributes, otherwise they won't be
displayed and it will be just confusing. This also unlocks further
refactoring.

* Move id to code scope conversion to command

Since it's the only place where it's used to it's more explicit.
Otherwise the newcomer to the code could think that it's used on the
view as well.

* Some delegations for uniformity

* Add scope error transformation

* Ensure document_scope_id is properly validated

* Extract the API connector instance to a method

I'll be reusing it inside commands.

* API validations in verification step

* Pass person_proxy to commands

For less indirection and less API calls.

* Indentation fix

* Log API errors

* Log connection refused as well

This is a classic in development when you forget to boot census.

* Bubble API errors app as a standard callout

* Add basic membership level step error handling

Required renaming the form attribute because of a rectify bug where the
parameter cannot be named like the form itself.

* Regenerate & cleanup all cassettes

* Simplify API module

* Remove unused hash key

* Split exception case

The interface of exceptions and http responses is not the same :S

* Catch up crowdfundings w/ latest connector changes

* Rename block params to better represent content

* Tiny refactor

* Compatibility with i18n friendly census API

* Extract error manipulation to a class

* Move `files` down to the form

* I18n compatibility for verification form errors

* Reduce indirection when grabbing user from command

* Prefer hash_including to regexp

* Fix faraday warning in debug mode

```
WARNING: Unexpected middleware set after the adapter. This won't be supported from Faraday 1.0.
```

It was also preventing the debug mode from working correctly.

* Add a bunch of missing tests

* Whitelist json body response codes

For everything else, return the code and log the response.

* Yet some more data command tests

* Limit scope queries to current organization