Checkout the code on coverage upload
This shouldn't be necessary but I'm guessing we can't ensure /app
exists unless we clone the code. I'd rather save the artifacts in
another folder so I don't need to checkout the whole project just to
submit coverage info to codeclimate but I want to confirm that this is
indeed the problem first.