Configure code coverage (#82)
* Start getting coverage info locally

* Coverage info aggregation

* Typo

* Not sure why the error, try a user-local path

* Debug base folder existance

* Create tmp folder before writing to it

* Make sure reporter is available for all jobs

* Remove incorrect folder switching

* Reduce phases by one

Too much artifacts to propagate otherwise. I think like this is good.

* Add missing working_directory config to main job

* Maybe this needs setting the working directory too?

* Checkout the code on coverage upload

This shouldn't be necessary but I'm guessing we can't ensure /app
exists unless we clone the code. I'd rather save the artifacts in
another folder so I don't need to checkout the whole project just to
submit coverage info to codeclimate but I want to confirm that this is
indeed the problem first.

* Yeah, that's most likely it

Pending improving it, let's confirm it 100%.

* Missing blank line

* Forgot to persist partial coverage reports

* Remove debugging info

* Try using global tmp

So I don't need to clone the whole repo when uploading coverage.

* Remove unnecessary configs

That's the working directory of the base image already.